+++ dev-tools/src/saf-docs-cli/generate.ts
@@ -1,0 +1,10 @@
+import { type MonorepoContext } from "../workspace.ts";
+import { generateTypeDoc } from "./generate-typedoc.ts";
+import { generateCliDocs } from "./generate-cli.ts";
+import { generateEnvDocs } from "./generate-env.ts";
+
+export const generateCommand = (monorepoContext: MonorepoContext) => {
+  generateTypeDoc(monorepoContext);
+  generateCliDocs(monorepoContext);
+  generateEnvDocs(monorepoContext);
+};
